refactor: move imagesource folder
