+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/ImageSource/ImageSource.stories.tsx
@@ -1,5 +1,5 @@
 import { Story, Meta } from '@storybook/react'
-import { journeysAdminConfig } from '../../../../../../../libs/storybook'
+import { journeysAdminConfig } from '../../../../../../libs/storybook'
 
 import { ImageSource } from './ImageSource'
